changing all floats by double

+++ src/layers/Activation.cc
@@ -10,7 +10,7 @@
 Activation::Activation(std::string functionName, int inputDim) : function_name(std::move(functionName)),
                                                                         input_dim(inputDim) {}
 
-float *Activation::forward(float *input) {
+double *Activation::forward(double *input) {
     if (function_name != "softmax")
         return apply_fct(input, input_dim, function_name);
     return apply_softmax(input, input_dim);
