+++ src/layers/Conv2D.hh
@@ -13,14 +13,14 @@ class Conv2D : public Layer {
 private:
     int input_N;
     int input_M;
-    float* kernel;
+    double* kernel;
     int kernel_size;
     std::string padding;
 
 public:
-    Conv2D(int inputN, int inputM, float *kernel, int kernelSize, std::string padding);
+    Conv2D(int inputN, int inputM, double *kernel, int kernelSize, std::string padding);
 
-    float* forward(float* input) override;
+    double* forward(double* input) override;
 };
