+++ src/mat_op.cc
@@ -13,17 +13,17 @@ int main() {
         int MA = 5;
         int NB = 4;
         int MB = 5;
-        float A[20] =
+        double A[20] =
                 {3., 3., 2., 1., 0.,
                 0., 0., 1., 3., 1.,
                 3., 1., 2., 2., 3.,
                 2., 0., 0., 2., 2.};
-        float B[20] =
+        double B[20] =
                 {1., 1., 1., 1., 1.,
                  1., 1., 1., 1., 1.,
                  1., 1., 1., 1., 1.,
                  1., 1., 1., 1., 1.};
-        float* C = mat_add(A, B, NA, MA, NB, MB);
+        double* C = mat_add(A, B, NA, MA, NB, MB);
         for (int i=0; i<NA; i++) {
             for (int j = 0; j < MA; j++) {
                 std::cout << C[i * MA + j] << ' ';
@@ -37,14 +37,14 @@ int main() {
         int MA = 3;
         int NB = 3;
         int MB = 4;
-        float A[6] =
+        double A[6] =
                 {3., 3., 2.,
                  0., 0., 1.};
-        float B[12] =
+        double B[12] =
                 {1., 1., 1., 1.,
                  1., 1., 1., 1.,
                  1., 1., 1., 1.};
-        float* C = mat_mult(A, B, NA, MA, NB, MB);
+        double* C = mat_mult(A, B, NA, MA, NB, MB);
         for (int i=0; i<NA; i++) {
             for (int j = 0; j < MB; j++) {
                 std::cout << C[i * MB + j] << ' ';
@@ -55,9 +55,9 @@ int main() {
 
 
     if (type=="test_dense") {
-        float A[9] =
+        double A[9] =
                 {12., 12., 17., 10., 17., 19., 9., 6., 14.};
-        float W2[18] =
+        double W2[18] =
                 {1., 1.,
                  -1., -1.,
                  1., 1.,
@@ -67,9 +67,9 @@ int main() {
                  1., 1.,
                  -1., -1.,
                  1., 1.};
-        float B2[2] = {1., 1.};
-        float* C = mat_mult(A, W2, 1, 9, 9, 2);
-        float* res = mat_add(C, B2, 1, 2, 1, 2);
+        double B2[2] = {1., 1.};
+        double* C = mat_mult(A, W2, 1, 9, 9, 2);
+        double* res = mat_add(C, B2, 1, 2, 1, 2);
         for (int i=0; i<1; i++) {
             for (int j = 0; j < 2; j++) {
                 std::cout << res[i * 2 + j] << ' ';
@@ -84,16 +84,16 @@ int main() {
         int NA = 4;
         int MA = 5;
         int KN = 3;
-        float A[20] =
+        double A[20] =
                 {3., 3., 2., 1., 0.,
                  0., 0., 1., 3., 1.,
                  3., 1., 2., 2., 3.,
                  2., 0., 0., 2., 2.};
 
-        float K[9] = {0., 1., 2.,
+        double K[9] = {0., 1., 2.,
                       2., 2., 0.,
                       0., 1., 2.};
-        float* C = mat_conv(A, K, NA, MA, KN, "valid");
+        double* C = mat_conv(A, K, NA, MA, KN, "valid");
         for (int i=0; i<NA - 2*(KN/2); i++) {
             for (int j = 0; j < MA - 2*(KN/2); j++) {
                 std::cout << C[i * (MA - 2*(KN/2)) + j] << ' ';
@@ -108,16 +108,16 @@ int main() {
         int NA = 4;
         int MA = 5;
         int KN = 3;
-        float A[20] =
+        double A[20] =
                 {1., 1., 1., 1., 1.,
                  1., 1., 1., 1., 1.,
                  1., 1., 1., 1., 1.,
                  1., 1., 1., 1., 1.};
 
-        float K[9] = {1., 10., 100.,
+        double K[9] = {1., 10., 100.,
                       1000., 10000., 100000.,
                       1000000., 10000000., 100000000.};
-        float* C = mat_conv(A, K, NA, MA, KN, "same");
+        double* C = mat_conv(A, K, NA, MA, KN, "same");
         for (int i=0; i<NA; i++) {
             for (int j = 0; j < MA; j++) {
                 std::cout << C[i * MA + j] << ' ';
@@ -129,8 +129,8 @@ int main() {
 
     if (type=="func") {
         int NA = 4;
-        float A[4] = {1., -1., 1., 1.};
-        float* C = apply_fct(A, NA, "exp");
+        double A[4] = {1., -1., 1., 1.};
+        double* C = apply_fct(A, NA, "exp");
         for (int i=0; i<NA; i++) {
             std::cout << C[i] << ' ';
         }
@@ -138,8 +138,8 @@ int main() {
 
     if (type=="softmax") {
         int NA = 4;
-        float A[4] = {1., -1., 1., 1.};
-        float* C = apply_softmax(A, NA);
+        double A[4] = {1., -1., 1., 1.};
+        double* C = apply_softmax(A, NA);
         for (int i=0; i<NA; i++) {
             std::cout << C[i] << ' ';
         }
