+++ src/matrices_operations/matrix_mult.hh
@@ -10,6 +10,6 @@
 #include <cmath>
 #include <cuda_runtime.h>
 
-float* mat_mult(float* A, float* B, int NA, int MA, int NB, int MB);
+double* mat_mult(double* A, double* B, int NA, int MA, int NB, int MB);
 
 #endif //GPGPU_MATRIX_MULT_HH
